Split language module service/components to separate files refs #735
